Handle AT-specific field mappings in toArray()
This allows the same logic to be applied regardless of which writer is used. For now it doesn't make much difference as there's only really a CsvWriter (as the FtpWriter uses CsvWriter as a base), but in the future we could end up creating other writers that will almost certainly need to have access to the same conversion logic. Having the mappings handled in toArray() also just makes more sense.